third_party/WebKit/LayoutTests/imagecapture/MediaStreamTrack-applyConstraints-getSettings.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/imagecapture-helpers.js"></script>
<script src="resources/mock-imagecapture.js"></script>
<body>
<canvas id='canvas' width=10 height=10/>
</body>
<script>

// This test verifies that the |constraints| configured in the mock Mojo
// service implementation, are returned by MediaStreamTrack.getSettings().

async_test(function(t) {
  var canvas = document.getElementById('canvas');
  var context = canvas.getContext('2d');
  context.fillStyle = 'red';
  context.fillRect(0, 0, 10, 10);

  const constraints = { advanced : [{ whiteBalanceMode : 'single-shot',
                                      exposureMode     : 'continuous',
                                      focusMode        : 'single-shot',

                                      pointsOfInterest : [{x : 0.1, y : 0.2},
                                                          {x : 0.3, y : 0.4}],

                                      exposureCompensation : 133.77,
                                      colorTemperature     : 6000,
                                      iso                  : 120.0,

                                      brightness           : 3,
                                      contrast             : 4,
                                      saturation           : 5,
                                      sharpness            : 6,

                                      zoom                 : 3.141592,

                                      torch                : true
                                    }]};
  var theMock = null;
  var videoTrack = null;
  mockImageCaptureReady
    .then(mock => {
      theMock = mock;
    },
    error => {
      assert_unreached('Error creating MockImageCapture: ' + error);
    })
    .then(() => {
      // |stream| must be created _after_ |mock| is constructed to give the
      // latter time to override the bindings.
      var stream = canvas.captureStream();
      videoTrack = stream.getVideoTracks()[0];

      // |videoTrack|'s capabilities gathering, just like the actual capture, is
      // a process kicked off right after creation, we introduce a small delay
      // to allow for those to be collected, since they are needed to understand
      // which constraints are supported in applyConstraints().
      // TODO(mcasas): this shouldn't be needed, https://crbug.com/711524.
      return new Promise(resolve => setTimeout(resolve, 100));
    })
    .then(() => videoTrack.applyConstraints(constraints))
    .then(appliedConstraints => {

      settings = videoTrack.getSettings();
      assert_equals(typeof settings, 'object');

      assert_equals(constraints.advanced[0].whiteBalanceMode,
                    settings.whiteBalanceMode, 'whiteBalanceMode');
      assert_equals(constraints.advanced[0].exposureMode, settings.exposureMode,
                    'exposureMode');
      assert_equals(constraints.advanced[0].focusMode, settings.focusMode,
                    'focusMode');

      assert_point2d_array_approx_equals(
          constraints.advanced[0].pointsOfInterest, settings.pointsOfInterest,
          0.01);

      assert_equals(constraints.advanced[0].exposureCompensation,
                    settings.exposureCompensation, 'exposureCompensation');
      assert_equals(constraints.advanced[0].colorTemperature,
                    settings.colorTemperature, 'colorTemperature');
      assert_equals(constraints.advanced[0].iso, settings.iso, 'iso');

      assert_equals(constraints.advanced[0].brightness, settings.brightness,
                    'brightness');
      assert_equals(constraints.advanced[0].contrast, settings.contrast,
                    'contrast');
      assert_equals(constraints.advanced[0].saturation, settings.saturation,
                    'saturation');
      assert_equals(constraints.advanced[0].sharpness, settings.sharpness,
                    'sharpness');

      assert_equals(constraints.advanced[0].zoom, settings.zoom, 'zoom');

      assert_equals(constraints.advanced[0].torch, settings.torch, 'torch');

      t.done();
    })
    .catch(error => {
      assert_unreached('Error applying constraints: ' + error.message);
    });

}, 'exercises an applyConstraints() - getSettings() cycle');

</script>
